requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
tqdm>=4.65.0
Pillow>=9.0.0
numpy>=1.24.0
pandas>=2.0.0
kagglehub>=0.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tensorboard>=2.13.0
tensorflow>=2.12.0
tensorflowjs>=4.10.0
onnx>=1.14.0
onnx-tf>=1.14.0
